+++ src/common/definitions/field-status.def.js
@@ -1,12 +1,12 @@
 /*
- * pwix:forms/src/common/definitions/check-status.def.js
+ * pwix:forms/src/common/definitions/field-status.def.js
  */
 
 import _ from 'lodash';
 
 import { pwixI18n } from 'meteor/pwix:i18n';
 
-export const CheckStatus = {
+export const FieldStatus = {
     C: {
         INVALID:    'INVALID',
         NONE:       'NONE',
@@ -36,35 +36,35 @@ export const CheckStatus = {
         }
     }
 };
-_.merge( CheckStatus, {
+_.merge( FieldStatus, {
     Order: [
-        CheckStatus.C.NONE,
-        CheckStatus.C.VALID,
-        CheckStatus.C.UNCOMPLETE,
-        CheckStatus.C.INVALID
+        FieldStatus.C.NONE,
+        FieldStatus.C.VALID,
+        FieldStatus.C.UNCOMPLETE,
+        FieldStatus.C.INVALID
     ],
 
     // check that the type is known
     _byType( type ){
-        if( !Object.keys( CheckStatus.K ).includes( type )){
-            console.warn( 'CheckStatus: unknown type', type );
+        if( !Object.keys( FieldStatus.K ).includes( type )){
+            console.warn( 'FieldStatus: unknown type', type );
             return null;
         }
-        return CheckStatus.K[type];
+        return FieldStatus.K[type];
     },
 
     /**
      * @returns {Array} the list of defined check types
      */
     Knowns(){
-        return Object.keys( CheckStatus.K );
+        return Object.keys( FieldStatus.K );
     },
 
     /**
      * @returns {String} the classes associated with this type
      */
     classes( type ){
-        const o = CheckStatus._byType( type );
+        const o = FieldStatus._byType( type );
         return o ? o.class : null;
     },
 
@@ -72,17 +72,17 @@ _.merge( CheckStatus, {
      * @returns {String} the name of the icon associated with this type
      */
     icon( type ){
-        const o = CheckStatus._byType( type );
+        const o = FieldStatus._byType( type );
         return o ? o.icon : '';
     },
 
     /**
-     * @param {CheckStatus} status
+     * @param {FieldStatus} status
      * @returns {Integer} the index of the status in the 'Order' array
      */
     index( status ){
-        for( let i=0 ; i<CheckStatus.Order.length ; ++i ){
-            if( CheckStatus.Order[i] === status ){
+        for( let i=0 ; i<FieldStatus.Order.length ; ++i ){
+            if( FieldStatus.Order[i] === status ){
                 return i;
             }
         }
@@ -94,22 +94,22 @@ _.merge( CheckStatus, {
      * @returns {String} the title associated with this type
      */
     title( type ){
-        const o = CheckStatus._byType( type );
+        const o = FieldStatus._byType( type );
         return o && o.title ? pwixI18n.label( I18N, o.title ) : '';
     },
 
     /**
-     * @param {Array<CheckStatus>} array a list of statuses
-     * @returns {CheckStatus} the worst statuses
+     * @param {Array<FieldStatus>} array a list of statuses
+     * @returns {FieldStatus} the worst statuses
      */
     worst( statuses ){
         let max = -1;
         statuses.forEach(( st ) => {
-            const index = CheckStatus.index( st );
+            const index = FieldStatus.index( st );
             if( index > max ){
                 max = index;
             }
         });
-        return CheckStatus.Order[max];
+        return FieldStatus.Order[max];
     }
 });